Move all login logic into LoginManager
Motivation: to keep everything in one place & for easier unit testing.
